moved OsmoGridModel to proper package
